:ok_hand: add trackPageActivities.spec.ts test file
